+++ examples/Features.tsx
@@ -47,7 +47,11 @@ export default function Features(): JSX.Element {
                 {/* Without any type, the features will be assumed to be a of a generic Geometry type */}
                 <RLayerVector
                     zIndex={5}
-                    /* This layer will be getting its data from an URL */
+                    /**
+                     * This layer will be getting its data from an URL, do not forget that in
+                     * OpenLayers 9.2 the format parsers now return RenderFeature by default unless
+                     * featureClass is explicitly specified
+                     */
                     format={new GeoJSON({featureProjection: 'EPSG:3857', featureClass: Feature})}
                     url='https://raw.githubusercontent.com/gregoiredavid/france-geojson/master/departements.geojson'
                     onPointerEnter={useCallback(