+++ src/layer/RLayerVector.tsx
@@ -14,11 +14,34 @@ import {Geometry} from 'ol/geom';
 /**
  * A vector layer
  *
- * Supports loading of features from external sources
+ * Supports loading of features from external sources.
  *
- * Requires an `RMap` context
+ * Requires an `RMap` context.
  *
- * Provides a vector layer context for JSX-declared `RFeature`s
+ * Provides a vector layer context for JSX-declared `RFeature`s.
+ *
+ * @example
+ * <RLayerVector>
+ *   <RFeature geometry={geom1} />
+ *   <RFeature geometry={geom2} />
+ * </RLayerVector>
+ *
+ * Features can also be typed when using TSX:
+ *
+ * @example
+ * <RLayerVector<Feature<Point>>
+ *   <RFeature geometry={point1} />
+ *   <RFeature geometry={point1} />
+ * </RLayerVector>
+ *
+ * In this case, all callbacks will be automatically typed.
+ *
+ * Since 3.0, RLayerVector also supports OpenLayers light-weight RenderFeatures.
+ * Currently these light-weight feature cannot be created using TSX/JSX and must
+ * be read from a file.
+ *
+ * @example
+ * <RLayerVector<RenderFeature> format={new GeoJSON({featureProjection: 'EPSG:3857'})} />
  */
 export default class RLayerVector<
     F extends FeatureLike = Feature<Geometry>